fix: particles peformance fix in swap page
